+++ sqlite/changelog.md
@@ -4,9 +4,9 @@
 * 2022/09/27
   * SQLiteを開く機能
   * SQLiteを閉じる機能
+  * 変数の型からSQLiteのTableを作成する機能
 
 # To Do
-* 変数の型からSQLiteのTableを作成する機能
 * SQLiteのTableからデータを読み取る機能
 * SQLiteのTableから検索する機能(オプションあり)
 * SQLiteのTableからデータを追加する機能